refactor: 댓글 수정 로직 분리
