+++ src/features/edit-comment/index.ts
@@ -1,3 +1,2 @@
-export { useCommentEditor } from "./model/edit-comment.hook";
-export { useNewCommentForm, useEditCommentDialog } from "./model/edit-comment.form";
-export type { NewCommentDraft } from "./model/edit-comment.form";
+export { useCommentEditor, useNewCommentForm, useEditCommentDialog } from "./model/edit-comment.hook";
+export type { NewCommentDraft } from "./model/edit-comment.hook";
